CPU Implementation is ready

+++ CPU Implementation/validation.c
@@ -1,6 +1,6 @@
 /**********************************************************************
  *
- * validation.cu (.c) -- Validation function for the number of triangles 
+ * validation.c -- Validation function for the number of triangles 
  *                    calculated
  *
  * Michail Iason Pavlidis <michailpg@ece.auth.gr>
@@ -13,7 +13,7 @@
 #include <math.h>
 #include "validation.h"
 
-int validation(long nT, long nT_Mat)
+int validation(int nT, int nT_Mat)
 {
 
   if ( nT != nT_Mat )
