+++ Sparse/validation.cu
@@ -13,12 +13,12 @@
 #include <math.h>
 #include "validation.h"
 
-int validation(long nT, long nT_Mat)
+int validation(int nT, int nT_Mat)
 {
 
   if ( nT != nT_Mat )
   {
-    printf("Validation FAILED: nT = %ld, while correct value nT_Mat = %ld \n", nT, nT_Mat);
+    printf("Validation FAILED: nT = %d, while correct value nT_Mat = %d \n", nT, nT_Mat);
     return 0;
   }
